==> notebooks/03-Filtrado-de-datods.ipynb <==
# Librerias y dependecias

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import pandas as pd
from ytmusicapi import YTMusic
import re
from tqdm import tqdm
import time
import random
import requests
import os
from dotenv import load_dotenv

# Autenticación

load_dotenv() # carga .env en process env
CLIENT_ID = os.getenv("CLIENT_ID")
CLIENT_SECRET = os.getenv("CLIENT_SECRET")
sp = spotipy.Spotify(
    auth_manager=SpotifyClientCredentials(
        client_id=CLIENT_ID,
        client_secret=CLIENT_SECRET
    )
)
yt = YTMusic()

# Filtrado de resultados TOP 100

#Carga del dataset
df_final = pd.read_csv("../data/interim/03_dataset_con_audiofeatures_reccobeats.csv", encoding="utf-8", low_memory=False)


df_final.info()

print(df_final["genero"].value_counts())

df_final.info()

# Eliminar registros con nulos en acousticness
df_filtrado = df_final.dropna(subset=["acousticness"]).copy()

# Asegurar que popularidad sea numérica
df_filtrado["popularidad"] = pd.to_numeric(df_filtrado["popularidad"], errors="coerce")

# Rellenar popularidad faltante con promedio por género
df_filtrado["popularidad"] = df_filtrado.groupby("genero")["popularidad"].transform(
    lambda x: x.fillna(x.mean())
)

# Función auxiliar: top N por género
def top_por_genero(df, genero, n=100):
    subset = df[df["genero"].str.lower() == genero.lower()]
    subset_top = subset.sort_values("popularidad", ascending=False).head(n)
    return subset_top

# Definir los géneros a limitar
generos_limitados = ["rock", "kpop", "regueton","hiphop","pop"]

# Filtrar el resto (que no se tocan)
otros_generos = df_filtrado[~df_filtrado["genero"].str.lower().isin(generos_limitados)]

# Tomar top 100 de los géneros seleccionados
top_rock = top_por_genero(df_filtrado, "rock", 150)
top_kpop = top_por_genero(df_filtrado, "kpop", 70)
top_regueton = top_por_genero(df_filtrado, "regueton", 100)
top_hiphop = top_por_genero(df_filtrado, "hiphop", 150)
top_pop = top_por_genero(df_filtrado, "pop", 50)

# Unir todo: los tops + los géneros restantes sin cambios
df_final_filtrado = pd.concat([otros_generos, top_rock, top_kpop, top_regueton,top_hiphop,top_pop], ignore_index=True)

# Resumen
print(df_final_filtrado["genero"].value_counts())
print(f"✅ Dataset final: {df_final_filtrado.shape[0]} filas")

# Guardar
df_final_filtrado.to_csv("../data/processed/04_dataset_filtrado.csv", index=False)
